Adding index elements for more exercises
I'm having a hell of a time making sure I have everything
this list. Since there were challenges and problem sets
and exercises, I initially searched specifically for things
that had a solutions lecture. But now I realize that are
a bunch of exercises that don't have one.

At this point I'm just going to be done with this for now
and get back to work. If there are any more I missed, I'll
find them when I get to them.